Refactor codebase to use Decimal for financial calculations
- Migrated models, engine, and strategies to use Decimal.
- Added Precision class for handling rounding and SDK conversion.
- Updated all tests to use Decimal and converted to unittest.
- Ensured type safety with mypy.

+++ src/strategy/base.py
@@ -3,10 +3,11 @@ from src.engine.context import StrategyContext
 from src.model import OrderFill, Cloid, OrderFailure
 from src.strategy.types import StrategySummary, GridState
 from typing import Union
+from decimal import Decimal
 
 class Strategy(ABC):
     @abstractmethod
-    def on_tick(self, price: float, ctx: StrategyContext):
+    def on_tick(self, price: Decimal, ctx: StrategyContext):
         pass
 
     @abstractmethod
